+++ src/core/interfaces/base-components-props.ts
@@ -1,6 +1,11 @@
 interface BaseComponentProps {
   style?: string;
   hover?: string;
+  bgImage?: string;
+  background?: string;
+  bgPos?: string;
+  bgSize?: string;
+  bgRepeat?: string;
 }
 
 export default BaseComponentProps;